# [yumemi] Improve speed by using HashMap instead of Vec in GameScoreSummary